move custom covers admin to own expansion-panel and add docs
