+++ frontend/src/components/admin/tabs/library-tab.vue
@@ -6,232 +6,70 @@
         :inputs="AdminLibraryCreateUpdateInputs"
       />
     </header>
-    <AdminTable
-      :headers="headers"
-      :items="items"
+    <AdminLibraryTable
+      :items="libraryItems"
       :sort-by="[{ key: 'path', order: 'asc' }]"
-    >
-      <template #no-data>
-        <td class="adminNoData" colspan="100%">
-          Add a Library to start using Codex
-        </td>
-      </template>
-      <template #[`item.events`]="{ item }">
-        <v-checkbox-btn :model-value="item.events" disabled />
-      </template>
-      <template #[`item.poll`]="{ item }">
-        <v-checkbox-btn :model-value="item.poll" disabled />
-      </template>
-      <template #[`item.pollEvery`]="{ item }">
-        <span :class="{ disabled: !item.poll }">
-          {{ item.pollEvery }}
-        </span>
-      </template>
-      <template #[`item.lastPoll`]="{ item }">
-        <DateTimeColumn :dttm="item.lastPoll" />
-      </template>
-      <template #[`item.groups`]="{ item }">
-        <v-icon v-if="item.coversOnly" disabled>{{
-          mdiCircleOffOutline
-        }}</v-icon>
-        <RelationChips
-          v-else
-          :pks="item.groups"
-          :objs="groups"
-          group-type
-          title-key="name"
-        />
-      </template>
-      <template #[`item.actions`]="{ item }">
-        <ConfirmDialog
-          :icon="mdiDatabaseClockOutline"
-          :title-text="`Poll for updated ${item.label}s`"
-          :object-name="item.path"
-          :confirm-text="pollConfirmText(item)"
-          size="small"
-          density="compact"
-          @confirm="poll(item)"
-        />
-        <ConfirmDialog
-          :icon="mdiDatabaseSyncOutline"
-          :title-text="`Force update every ${item.label}`"
-          :object-name="item.path"
-          confirm-text="Force Update"
-          size="small"
-          density="compact"
-          @confirm="forcePoll(item)"
-        />
-        <AdminCreateUpdateDialog
-          table="Library"
-          :old-row="item"
-          :inputs="AdminLibraryCreateUpdateInputs"
-          :label="updateLabel(item)"
-          max-width="22em"
-          size="small"
-          density="compact"
-        />
-        <AdminDeleteRowDialog
-          v-if="!item.coversOnly"
-          table="Library"
-          :pk="item.pk"
-          :name="item.path"
-          size="small"
-          density="compact"
-        />
-      </template>
-    </AdminTable>
+    />
+
     <v-expand-transition>
-      <AdminFailedImportsPanel />
+      <AdminFailedImportsPanel id="failedImports" />
+    </v-expand-transition>
+
+    <v-expand-transition>
+      <AdminCustomCoversPanel id="customCovers" />
     </v-expand-transition>
   </div>
 </template>
 
 <script>
-import {
-  mdiCircleOffOutline,
-  mdiDatabaseClockOutline,
-  mdiDatabaseSyncOutline,
-  mdiOpenInNew,
-} from "@mdi/js";
 import { mapActions, mapState } from "pinia";
 import { markRaw } from "vue";
 
-import AdminCreateUpdateDialog from "@/components/admin/create-update-dialog/create-update-dialog.vue";
 import AdminLibraryCreateUpdateInputs from "@/components/admin/create-update-dialog/library-create-update-inputs.vue";
-import AdminTable from "@/components/admin/tabs/admin-table.vue";
-import DateTimeColumn from "@/components/admin/tabs/datetime-column.vue";
-import AdminDeleteRowDialog from "@/components/admin/tabs/delete-row-dialog.vue";
+import AdminCustomCoversPanel from "@/components/admin/tabs/custom-covers-panel.vue";
 import AdminFailedImportsPanel from "@/components/admin/tabs/failed-imports-panel.vue";
-import RelationChips from "@/components/admin/tabs/relation-chips.vue";
-import ConfirmDialog from "@/components/confirm-dialog.vue";
-import { getDateTime } from "@/datetime";
+import AdminLibraryTable from "@/components/admin/tabs/library-table.vue";
 import { useAdminStore } from "@/stores/admin";
-import { useBrowserStore } from "@/stores/browser";
-import { useCommonStore } from "@/stores/common";
 
 export default {
   name: "AdminLibrariesTab",
   components: {
-    AdminTable,
-    AdminDeleteRowDialog,
+    AdminLibraryTable,
     AdminFailedImportsPanel,
-    AdminCreateUpdateDialog,
-    RelationChips,
-    ConfirmDialog,
-    DateTimeColumn,
+    AdminCustomCoversPanel,
   },
   data() {
     return {
-      lastUpdate: {
-        pk: 0,
-        field: undefined,
-      },
-      mdiCircleOffOutline,
-      mdiDatabaseClockOutline,
-      mdiDatabaseSyncOutline,
-      mdiOpenInNew,
       AdminLibraryCreateUpdateInputs: markRaw(AdminLibraryCreateUpdateInputs),
-      headers: [
-        { title: "Path", key: "path", align: "start" },
-        {
-          title: "Watch File Events",
-          key: "events",
-        },
-        {
-          title: "Poll Files Periodically",
-          key: "poll",
-        },
-        { title: "Poll Every", key: "pollEvery" },
-        { title: "Last Poll", key: "lastPoll" },
-        { title: "Groups", key: "groups" },
-        { title: "Actions", key: "actions", sortable: false },
-      ],
     };
   },
   computed: {
     ...mapState(useAdminStore, {
-      groups: (state) => state.groups,
-      // libraries: (state) => state.libraries,
-      // customCoversDirs: (state) => state.customCoversDirs,
-      items(state) {
+      libraryItems(state) {
         const annotatedItems = [];
         for (const library of state.libraries) {
-          const annotatedItem = { ...library };
-          annotatedItem.label = annotatedItem.coversOnly
-            ? "custom group cover"
-            : "comic";
-          annotatedItems.push(annotatedItem);
+          if (!library.coversOnly) {
+            const annotatedItem = { ...library };
+            annotatedItem.label = annotatedItem.coversOnly
+              ? "custom group cover"
+              : "comic";
+            annotatedItems.push(annotatedItem);
+          }
         }
         return annotatedItems;
       },
-    }),
-    ...mapState(useCommonStore, {
-      formErrors: (state) => state.form?.errors,
-    }),
-    ...mapState(useBrowserStore, {
-      twentyFourHourTime: (state) => state.settings.twentyFourHourTime,
     }),
   },
   mounted() {
     this.loadTables(["Group", "Library", "FailedImport"]);
   },
   methods: {
-    ...mapActions(useAdminStore, [
-      "updateRow",
-      "clearErrors",
-      "librarianTask",
-      "loadTables",
-    ]),
-    formatDateTime: (dttm) => {
-      return dttm ? getDateTime(dttm, this.twentyFourHourTime) : "";
-    },
-    changeCol(pk, field, val) {
-      this.lastUpdate.pk = pk;
-      this.lastUpdate.field = field;
-      const data = { [field]: val };
-      this.updateRow("Library", pk, data);
-    },
-    getFormErrors(pk, field) {
-      if (pk === this.lastUpdate.pk && field === this.lastUpdate.field) {
-        return this.formErrors;
-      }
-    },
-    libraryLabel(item, long = true) {
-      let label = "";
-      if (item.coversOnly) {
-        if (long) {
-          label += "Custom Group ";
-        }
-        label += "Cover Dir";
-      } else {
-        label += "Library";
-      }
-      return label;
-    },
-    poll(item) {
-      const label = this.libraryLabel(item);
-      this.librarianTask("poll", `Poll ${label} ${item.pk}`, item.pk);
-    },
-    forcePoll(item) {
-      const label = this.libraryLabel(item);
-      this.librarianTask(
-        "poll_force",
-        `Force Poll ${label} ${item.pk}`,
-        item.pk,
-      );
-    },
-    pollConfirmText(item) {
-      const label = this.libraryLabel(item, false);
-      return `Poll ${label}`;
-    },
-    updateLabel(item) {
-      return item.coversOnly ? "Cover Dir" : "";
-    },
+    ...mapActions(useAdminStore, ["loadTables"]),
   },
 };
 </script>
 <style scoped lang="scss">
-.disabled {
-  color: rgb(var(--v-theme-textDisabled)) !important;
+#customCovers {
+  margin-top: 60px;
 }
 </style>
